migrate train-ticket sources to 3.2.0 (completes 070908f4)

+++ applications/train-ticket/src/main/java/pt/ulisboa/tecnico/socialsoftware/trainticket/events/OrderContactsUpdatedEvent.java
@@ -1,12 +1,12 @@
 package pt.ulisboa.tecnico.socialsoftware.trainticket.events;
 
 import jakarta.persistence.Entity;
-import pt.ulisboa.tecnico.socialsoftware.ms.notification.Event;
+import pt.ulisboa.tecnico.socialsoftware.ms.aggregate.Event;
 
 @Entity
 public class OrderContactsUpdatedEvent extends Event {
     private Integer contactsAggregateId;
-    private Integer contactsVersion;
+    private Long contactsVersion;
     private String contactsName;
     private String contactsDocumentNumber;
 
@@ -18,7 +18,7 @@ public class OrderContactsUpdatedEvent extends Event {
         super(aggregateId);
     }
 
-    public OrderContactsUpdatedEvent(Integer aggregateId, Integer contactsAggregateId, Integer contactsVersion, String contactsName, String contactsDocumentNumber) {
+    public OrderContactsUpdatedEvent(Integer aggregateId, Integer contactsAggregateId, Long contactsVersion, String contactsName, String contactsDocumentNumber) {
         super(aggregateId);
         setContactsAggregateId(contactsAggregateId);
         setContactsVersion(contactsVersion);
@@ -34,11 +34,11 @@ public class OrderContactsUpdatedEvent extends Event {
         this.contactsAggregateId = contactsAggregateId;
     }
 
-    public Integer getContactsVersion() {
+    public Long getContactsVersion() {
         return contactsVersion;
     }
 
-    public void setContactsVersion(Integer contactsVersion) {
+    public void setContactsVersion(Long contactsVersion) {
         this.contactsVersion = contactsVersion;
     }
