+++ applications/train-ticket/src/main/java/pt/ulisboa/tecnico/socialsoftware/trainticket/events/PriceConfigTrainUpdatedEvent.java
@@ -1,12 +1,12 @@
 package pt.ulisboa.tecnico.socialsoftware.trainticket.events;
 
 import jakarta.persistence.Entity;
-import pt.ulisboa.tecnico.socialsoftware.ms.notification.Event;
+import pt.ulisboa.tecnico.socialsoftware.ms.aggregate.Event;
 
 @Entity
 public class PriceConfigTrainUpdatedEvent extends Event {
     private Integer trainAggregateId;
-    private Integer trainVersion;
+    private Long trainVersion;
     private String trainTypeName;
 
     public PriceConfigTrainUpdatedEvent() {
@@ -17,7 +17,7 @@ public class PriceConfigTrainUpdatedEvent extends Event {
         super(aggregateId);
     }
 
-    public PriceConfigTrainUpdatedEvent(Integer aggregateId, Integer trainAggregateId, Integer trainVersion, String trainTypeName) {
+    public PriceConfigTrainUpdatedEvent(Integer aggregateId, Integer trainAggregateId, Long trainVersion, String trainTypeName) {
         super(aggregateId);
         setTrainAggregateId(trainAggregateId);
         setTrainVersion(trainVersion);
@@ -32,11 +32,11 @@ public class PriceConfigTrainUpdatedEvent extends Event {
         this.trainAggregateId = trainAggregateId;
     }
 
-    public Integer getTrainVersion() {
+    public Long getTrainVersion() {
         return trainVersion;
     }
 
-    public void setTrainVersion(Integer trainVersion) {
+    public void setTrainVersion(Long trainVersion) {
         this.trainVersion = trainVersion;
     }
