+++ applications/train-ticket/src/main/java/pt/ulisboa/tecnico/socialsoftware/trainticket/microservices/route/aggregate/RouteStation.java
@@ -5,7 +5,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToOne;
 
-import pt.ulisboa.tecnico.socialsoftware.ms.domain.aggregate.Aggregate.AggregateState;
+import pt.ulisboa.tecnico.socialsoftware.ms.aggregate.Aggregate.AggregateState;
 
 import pt.ulisboa.tecnico.socialsoftware.trainticket.shared.dtos.RouteStationDto;
 import pt.ulisboa.tecnico.socialsoftware.trainticket.shared.dtos.StationDto;
@@ -19,7 +19,7 @@ public class RouteStation {
     private Integer distanceFromStart;
     private String stationName;
     private Integer stationAggregateId;
-    private Integer stationVersion;
+    private Long stationVersion;
     private AggregateState stationState;
     @OneToOne
     private Route route;
@@ -92,11 +92,11 @@ public class RouteStation {
         this.stationAggregateId = stationAggregateId;
     }
 
-    public Integer getStationVersion() {
+    public Long getStationVersion() {
         return stationVersion;
     }
 
-    public void setStationVersion(Integer stationVersion) {
+    public void setStationVersion(Long stationVersion) {
         this.stationVersion = stationVersion;
     }
